# server/openapi.yaml
openapi: 3.0.3
info:
  title: Metro Verify API
  description: >
    Digital Legal Metrology Verification & Lifecycle Management Platform (SIH26036).
    Covers Owner, Administrator, and LMO/GATC Officer endpoints, plus the AI chatbot.
    Endpoints marked "(stored procedure)" call a MySQL stored procedure for an atomic,
    multi-step operation rather than a plain Prisma query.
  version: "1.0.0"
servers:
  - url: http://localhost:4000
    description: Local development backend
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BulkRequest:
      type: object
      properties:
        bulkRequestId: { type: string, example: "BR-001" }
        ownerId: { type: string }
        totalInstruments: { type: integer }
        status: { type: string, enum: [SUBMITTED, PENDING, IN_PROGRESS, COMPLETED] }
...
security:
  - bearerAuth: []
paths:
...
  /api/bulk-requests:
    get:
      tags: [Owner - Bulk Requests]
      summary: List the logged-in owner's bulk requests
      responses:
        "200":
          description: List of bulk requests
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BulkRequest" }
    post:
      tags: [Owner - Bulk Requests]
      summary: Submit a bulk verification request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category: { type: string }
                numberOfItems: { type: integer, example: 25 }
      responses:
        "201":
          description: Bulk request created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BulkRequest" }

  # ---------------- OWNER: INSTRUMENTS ----------------
...
  /api/bulk-requests/{id}/split:
    post:
      tags: [Admin - Allocation]
      summary: Split a bulk request into batches across officers (stored procedure — sp_SplitBulkBatch)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                officerAllocations:
                  type: array
                  items:
                    type: object
                    properties:
                      officerId: { type: string }
                      count: { type: integer }
      responses:
        "200": { description: Batch and per-officer sub-assignments created as one transaction }

  # ---------------- ADMIN: COMPLIANCE ----------------
...